backend/requirements.txt
flask
flask-cors
requests
gspread
oauth2client
gunicorn
groq
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib